+++ examples/copyConfigForTestsOnly/config/cloud.js
@@ -1,7 +1,7 @@
 module.exports = {
   name:       'PrivateCloud',
   type:       'standalone',
-  controller: '127.0.0.1',
+  controller: 'tcp://127.0.0.1',
   pubSubPort: 3000,
   reqResPort: 3001,
   health:     '2s'