+++ examples/impressCloudController/instanceC0/applications/example/app/access.js
@@ -1,0 +1,9 @@
+module.exports = {
+  guests: true,
+  logged: true,
+  http:   true,
+  https:  true,
+  intro:  false,
+  index:  true,
+  groups: []
+}