+++ examples/impressCloudController/instanceC0/applications/example/config/application.js
@@ -1,0 +1,7 @@
+// Application configuration
+
+module.exports = {
+  slowTime:    '1s',
+  // preload:     true, // preload application to memory at startup (default: false)
+  // allowOrigin: '*' // set HTTP header Access-Control-Allow-Origin (default: not set)
+};