+++ examples/impressCloudController/instanceC1/applications/example/config/example.js
@@ -1,0 +1,12 @@
+// Application specific configuration
+// You can place js file in this directory with your own structure and it will be loaded as an application config section
+
+module.exports = {
+
+  parameter: 'value',
+
+  subSection: {
+    parameter: 'value'
+  }
+
+};