+++ examples/impressCloudController/instanceC1/applications/example/init/ccEvant.js
@@ -1,0 +1,6 @@
+if (api.cluster.isWorker) {
+  setInterval(function() {
+    console.log('SEND');
+    application.events.sendGlobal('TestEvent', { test: "data" });
+  }, 5000);
+}