+++ examples/impressCloudController/instanceC2/applications/example/app/events/connect.sse/get.js
@@ -1,0 +1,6 @@
+module.exports = function(client, callback) {
+
+  client.eventChannel = 'TestEvent';
+  callback();
+
+}