+++ examples/impressCloudController/instanceC2/applications/example/app/events/sendEvent.json/get.js
@@ -1,0 +1,6 @@
+module.exports = function(client, callback) {
+
+  application.events.sendGlobal('TestEvent', { test: "data" });
+  callback("Ok");
+
+}