+++ examples/impressCloudController/instanceC2/applications/example/config/hosts.js
@@ -1,0 +1,9 @@
+// Hosts (and IPs) array to be handled by application
+// Wildcard '*' is allowed for masking random or empty substring
+
+module.exports = [
+  '127.0.0.1',        // IP address
+  // 'domain.com',    // domain name
+  // '*.domain.com',  // masked domain name
+  // '*",             // all
+];